Unify edit bracket interface
